chore

+++ wip/core/mask.ipp
@@ -13,22 +13,22 @@ public:
     size_t next() const        { return -1; }
 };
 
-template<size_t N> Mask<N>& Mask<N>::on(uint64_t index) {
+template<size_t N> Mask<N>& Mask<N>::on(size_t index) {
     flags[index >> 6] |= (1ull << uint64_t(index & (64 - 1))); // [index / 64] |= index % 64
     return *this;
 }
 
-template<size_t N> Mask<N>& Mask<N>::off(uint64_t index) {
+template<size_t N> Mask<N>& Mask<N>::off(size_t index) {
     flags[index >> 6] &= ~(1ull << uint64_t(index & (64 - 1))); // [index / 64] &= index % 64
     return *this;
 }
 
-template<size_t N> Mask<N>& Mask<N>::toggle(uint64_t index) {
+template<size_t N> Mask<N>& Mask<N>::toggle(size_t index) {
     flags[index >> 6] ^= (1ull << uint64_t(index & (64 - 1))); // [index / 64] ^= index % 64
     return *this;
 }
 
-template<size_t N> bool Mask<N>::check(uint64_t index) const {
+template<size_t N> bool Mask<N>::check(size_t index) const {
     return (flags[index >> 6] >> uint64_t(index & (64 - 1))) & 1ull; // ([index / 64] >> index % 64) & 1
 }
